backend/requirements.txt
pymongo==4.3.3
dnspython==2.3.0
PyJWT==2.6.0
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.3.1
pytest-mock==3.10.0
boto3==1.26.129
botocore==1.29.129
mangum==0.17.0
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0
email-validator==2.0.0